add new input storybook
